photo_service/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
mysql-connector-python==8.0.26
Pillow==8.3.1